requirements.txt
streamlit
nltk
scikit-learn
pdfplumber
pytesseract
Pillow
